feat: Updates etl.transform.sherpa_romeo
Added a safety lock to avoid crashing with bad requests. Also added one
additional for loop, because I realized that I could have more data
being missed on the previous logic (and it proved to be the case). And
I created a function to selectivelly title-case the journals's names, to
 be applied before calling SHERPA/RoMEo, because I noticed their API is
also case sensitive.